requirements.txt
# Python 3.10+ recommandé
Pillow==9.5.0
opencv-python==4.11.0.86
numpy==1.24.4
tkinterdnd2==0.4.3
